issue fixed

+++ resources/views/backend/app/promoter-details.blade.php
@@ -157,9 +157,11 @@
                             ?>
                             <div class="col-md-12">
                                 <h5 class="card-title form-head pr-2">Management Information ({{isset($row->first_name) ? $i : '1'}})
-                                    @if(count($ownerDetails) > 1 && $appData->status == 0 || count($ownerDetails) > 1 && $appData->status > 0 && \Helpers::isChangeAppStatusAllowed($appData->curr_status_id))
-                                    <span class="float-right"><a href="javascript:void(0)" class=" text-danger" onclick="deleteManagementInfo({{ $row->biz_owner_id }})"><i class="fa fa-trash"></i></a></span>
-                                    @endif
+                                    @can('delete_management_info')
+                                        @if(count($ownerDetails) > 1 && $appData->status == 0 || count($ownerDetails) > 1 && $appData->status > 0 && \Helpers::isChangeAppStatusAllowed($appData->curr_status_id))
+                                        <span class="float-right"><a href="javascript:void(0)" class=" text-danger" onclick="deleteManagementInfo({{ $row->biz_owner_id }})"><i class="fa fa-trash"></i></a></span>
+                                        @endif
+                                    @endcan
                                 </h5>
 
                                 <div class="row">
